+++ CHANGES.rst
@@ -6,9 +6,10 @@ Unreleased
 
 **Added**
 
-- Added support for in ``code-block`` directive.
+- Added support for ``code-block`` directive.
 - Added support for ``highlight`` directive.
 - Added support for Sphinx metadata fields.
+- Added support for Sphinx Python directives.
 
 1.6.1 (2023/12/12)
 ------------------